requirements.txt
google-api-python-client==1.9.1
google-cloud-speech==1.3.2
google-cloud-storage==1.28.1
jupyter==1.0.0
pydub==0.24.1
pytest==5.4.3